Fix #138; Rename webpackConfig
webpack.config.babel => webpack.config
